requirements.txt
requests==2.32.5
selenium==4.40.0
pytest==9.0.2
allure-pytest==2.15.3

# В коде не должно быть специфичных данных — вынесите их в файл:
# настройки окружения (URL, пути до файлов и т. д.);
# тестовые данные (логины, пароли, токены и т. д.).
#  Что это значит?

# Все методы содержат аннотации типов для параметров и возвращаемого значения.

# Все ключевые действия и проверки в тестах обернуты в allure.step. Все тесты содержат декораторы @allure.title, @allure.story.

# Файл README.md содержит: описание проекта, четкие инструкции по запуску тестов (все режимы), ссылку на финальный проект по ручному тестированию.